refactor: add relative import to utils folder
